Ajustes

+++ WebProdutos/Models/Produto.cs
@@ -10,7 +10,7 @@ namespace WebProdutos.Models
         [Display(Name ="Código")]
         public int Id { get; set; }
         public string nome { get; set; }
-        public decimal valor { get; set; }
+        public float valor { get; set; }
         public int estoque { get; set; }
     }
 }
